% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_block.R
\name{make_block_grid}
\alias{make_block_grid}
\title{Write out traps from a block as a trapgrid file}
\usage{
make_block_grid(gridname = "footest", traps = traps, lambda = 0.005)
}
\arguments{
\item{gridname}{Name for the trapping grid file to be output}

\item{traps}{Coordinates of traps}

\item{lambda}{The trap efficiency}
}
\value{
A trapping grid text file
}
\description{
Write out traps from a block as a trapgrid file
}
