% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_block.R
\name{make_block_outbreak}
\alias{make_block_outbreak}
\title{Setup outbreak file from random blocks}
\usage{
make_block_outbreak(
  traps = traps,
  block = block,
  in_orchard = FALSE,
  nOutbreaks = 10,
  outbreak_name = "outbreaks",
  orchard_buf = 25,
  outbreak_buf = 250
)
}
\arguments{
\item{traps}{The two-column trap coordinates}

\item{block}{The polygon of the orchard/block boundary}

\item{in_orchard}{True/False flag for whether flies emerge from within OR outside orchard/block}

\item{nOutbreaks}{Number of outbreak locations}

\item{outbreak_name}{Name of file to write out}

\item{orchard_buf}{Distance around orchard (in metres) that outbreaks cannot occur}

\item{outbreak_buf}{Distance from orchard in which outbreaks can occur}
}
\value{
A trapping grid text file
}
\description{
Setup outbreak file from random blocks
}
\author{
Matt Hill
}
