% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{make_outbreak_file}
\alias{make_outbreak_file}
\title{Setup outbreak file}
\usage{
make_outbreak_file(
  traps = traps,
  in_orchard = TRUE,
  nOutbreaks = 10,
  outbreak_name = "outbreaks",
  lambda = 0.05,
  per_area = FALSE,
  orchard_buf = 25,
  outbreak_buf = 250
)
}
\arguments{
\item{traps}{Name for the trapping grid file to be output}

\item{in_orchard}{True/False flag for whether flies can emerge from within orchard}

\item{nOutbreaks}{Number of outbreaks}

\item{outbreak_name}{Name of file to write out}

\item{lambda}{The trap efficiency}

\item{per_area}{Flag to override the num_outbreaks and do it by area}

\item{orchard_buf}{Distance around orchard (in metres) that outbreaks cannot occur}

\item{outbreak_buf}{Distance around orchard that outbreaks can occur}
}
\value{
A trapping grid text file
}
\description{
Setup outbreak file
}
