% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_block.R
\name{make_random_block}
\alias{make_random_block}
\title{Create random and regular blocks}
\usage{
make_random_block(n.sides = 4, block.size = 1e+05, regular.block = F)
}
\arguments{
\item{n.sides}{Shape of the block, how many sides (3+)}

\item{block.size}{Area of the block in metres squared}

\item{regular.block}{Force block to have evenly sized and angled edges (e.g. regular shape)}
}
\value{
A SpatialPolygon of specified size and shape
}
\description{
Create random and regular blocks
}
\author{
Matt Hill
}
