% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{make_random_grid}
\alias{make_random_grid}
\title{Setup random trapping grid}
\usage{
make_random_grid(
  n.traps = 10,
  x1 = 2000,
  y1 = 2000,
  d = 400,
  trials = 1000,
  perim = FALSE,
  gridname = "footest",
  lambda = 0.02
)
}
\arguments{
\item{n.traps}{Number of random traps}

\item{x1}{Width of area (begins at 0)}

\item{y1}{Height of area (begins at 0)}

\item{d}{Minimum distance between each trap}

\item{trials}{How many attempts to generate trap arrangement with given parameters}

\item{perim}{Include perimeter traps or not}

\item{gridname}{Name for the trapping grid file to be output}

\item{lambda}{The trap efficiency}
}
\value{
A trapping grid text file
}
\description{
Generate a random distribution of grids on a defined rectangle (in metres)
Can inclue a perimeter trap on each boundary if desired
Calculates that each trap is a mininum distance apart before writing out trap grid file
}
