% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_block.R
\name{make_random_traps}
\alias{make_random_traps}
\title{Create random and regular traps}
\usage{
make_random_traps(
  block = block,
  block.size = 1e+05,
  n.traps = 4,
  regular.traps = F,
  min.dist = 100,
  longest.edge = F,
  perim = F
)
}
\arguments{
\item{block}{A SpatialPolygon - usually generated using \code{make_random_block()}}

\item{n.traps}{How many traps to place in block}

\item{regular.traps}{Place traps in a regular pattern through block}

\item{min.dist}{Minimum distance between all traps}

\item{longest.edge}{Identify the longest edge of the block and place a trap in the middle of it}

\item{perim}{Space n traps regularly around the perimeter}
}
\value{
Trap points, block polygon
}
\description{
Create random and regular traps
}
\author{
Matt Hill
}
