% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{make_regular_grid}
\alias{make_regular_grid}
\title{Setup regular trapping grid}
\usage{
make_regular_grid(
  gridname = "footest",
  gridSize = c(364, 572),
  gridSpace = 80,
  lambda = 0.02
)
}
\arguments{
\item{gridname}{Name for the trapping grid file to be output}

\item{gridSize}{Describes the bottom right corner in metres, from the top left corner which is 0,0}

\item{gridSpace}{The distance between regularly spaced traps}

\item{lambda}{The trap efficiency}
}
\value{
A trapping grid text file
}
\description{
Setup regular trapping grid
}
