% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{trapgridR}
\alias{trapgridR}
\title{Retrieve a Java instance of TrapGrid.}
\usage{
trapgridR(
  filepath = "inst/java/foogrid",
  nDays = 14,
  nFlies = 100,
  nSim = 10,
  D = 10^5
)
}
\arguments{
\item{filepath}{Path of the trapping grid file}
}
\value{
Model output
}
\description{
Retrieve a Java instance of TrapGrid
}
