% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traps_density.R
\name{traps_per_distance}
\alias{traps_per_distance}
\title{Random traps per distance from trap}
\usage{
traps_per_distance(points = points, poly = poly, dist.thresh = 100)
}
\arguments{
\item{points}{The two-column existing trap coordinates}

\item{poly}{The polygon of the orchard/block boundary}

\item{dist.thresh}{Minimum distance between all traps and all parts of the block}
}
\value{
2 column coordinates of all traps
}
\description{
Random traps per distance from trap
}
\author{
Matt Hill
}
