% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traps_density.R
\name{traps_per_hectare}
\alias{traps_per_hectare}
\title{Random traps per hectare}
\usage{
traps_per_hectare(points = points, poly = poly, traps.per.hectare = 1)
}
\arguments{
\item{points}{The two-column existing trap coordinates}

\item{poly}{The polygon of the orchard/block boundary}

\item{traps.per.hectare}{Density of traps per hectare required}
}
\value{
2 column coordinates of all traps
}
\description{
Random traps per hectare
}
\author{
Matt Hill
}
