% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{make_actual_grid}
\alias{make_actual_grid}
\title{Setup actual trapping grid}
\usage{
make_actual_grid(gridname = "footest", traps = traps, lambda = 0.005)
}
\arguments{
\item{gridname}{Name for the trapping grid file to be output}

\item{traps}{Trap coordinates}

\item{lambda}{The trap efficiency}
}
\value{
A trapping grid text file
}
\description{
Setup actual trapping grid
}
