% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{make_outbreak_file}
\alias{make_outbreak_file}
\title{Setup outbreak file}
\usage{
make_outbreak_file(
  traps = traps,
  in_orchard = TRUE,
  nSim = 10,
  outbreak_name = "outbreaks",
  lambda = 0.05,
  per_area = FALSE
)
}
\arguments{
\item{traps}{Name for the trapping grid file to be output}

\item{in_orchard}{True/False flag for whether flies can emerge from within orchard}

\item{nSim}{Number of sims}

\item{outbreak_name}{Name of file to write out}

\item{lambda}{The trap efficiency}

\item{per_area}{Flag to override the num_outbreaks and do it by area}
}
\value{
A trapping grid text file
}
\description{
Setup outbreak file
}
