% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{make_outbreak_file}
\alias{make_outbreak_file}
\title{Setup outbreak file}
\usage{
make_outbreak_file(
  traps = traps,
  in_orchard = TRUE,
  nSim = 10,
  outbreak_name = "outbreaks"
)
}
\arguments{
\item{traps}{Name for the trapping grid file to be output}

\item{in_orchard}{True/False flag for whether flies can emerge from within orchard}
}
\value{
A trapping grid text file
}
\description{
Setup outbreak file
}
