% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapgridR.R
\name{trapgridR}
\alias{trapgridR}
\title{Retrieve a Java instance of TrapGrid}
\usage{
trapgridR(
  filepath = paste0(system.file(package = "trapgridR"), "/java/foogrid"),
  nDays = 14,
  nFlies = 100,
  nSim = 10,
  D = 10^5,
  outbreaks = NULL,
  TGA = FALSE
)
}
\arguments{
\item{filepath}{Path of the trapping grid file}

\item{nDays}{Number of days to run simulation}

\item{nFlies}{Number of flies per outbreak}

\item{nSim}{Number of simulations}

\item{D}{Dispersal parameter}

\item{outbreaks}{Outbreaks file}
}
\value{
Model output
}
\description{
Retrieve a Java instance of TrapGrid
}
